Extract logic that is the same for every token to an abstract class
